# OpenAPI.yaml
openapi: 3.0.0
info:
  title: Bookstore API
  version: 1.0
  description: A simple RESTful API for managing books in a bookstore

paths:
  /books:
    get:
      summary: Retrieve all books
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  books:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        title:
                          type: string
                        author:
                          type: string
                        publishedYear:
                          type: integer

    post:
      summary: Add a new book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                author:
                  type: string
                publishedYear:
                  type: integer
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  '/books/{id}':
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
        description: ID of the book
    get:
      summary: Retrieve a specific book by ID
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  title:
                    type: string
                  author:
                    type: string
                  publishedYear:
                    type: integer
        '404':
          description: Book not found

    put:
      summary: Update a specific book by ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                author:
                  type: string
                publishedYear:
                  type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Book not found

    delete:
      summary: Delete a specific book by ID
      responses:
        '204':
          description: No Content
        '404':
          description: Book not found